cannot resize

+++ func/pic.py
@@ -20,11 +20,11 @@ def save_pic(event,pic_id) :
         img = Image.open('pic.jpg')
         img_size = img.size # (width, height)
         if img_size[0] > img_size[1] :
-            p_size = [240, img_size[1]*(240/img_size[0])]
-            new_size = [1024, img_size[1]*(1024/img_size[0])]
+            p_size = [240, int(img_size[1]*(240/img_size[0]))]
+            new_size = [1024, int(img_size[1]*(1024/img_size[0]))]
         else : 
-            p_size = [img_size[0]*(240/img_size[1]), 240]
-            new_size = [img_size[0]*(1024/img_size[1]), 1024]
+            p_size = [int(img_size[0]*(240/img_size[1])), 240]
+            new_size = [int(img_size[0]*(1024/img_size[1])), 1024]
         new_img= img.resize((p_size[0], p_size[1]),Image.ANTIALIAS)
         new_img.save('pic-p.jpg',quality=100)
         new_img = img.resize((1024, 650),Image.ANTIALIAS)
